Fix collision detection
Change the collision detection so it only detects a collision with the left side of a paddle when traveling right and the right side of a paddle when traveling left. Otherwise occasionally when the length of the vector is adjusted after collision the ball slows down in the x direction and will not far enough to be outside the paddle after the next dt triggering a chain reaction of collisions, direction changes and shortening x velocities.